+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/home/fabian/Documentos/Dev/dashboard-app/.eslintrc.js":"1"},{"size":898,"mtime":1652407608431,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","suppressedMessages":"6","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1foh4pi","/home/fabian/Documentos/Dev/dashboard-app/.eslintrc.js",[],[]]
