+++ package.json
@@ -17,7 +17,7 @@
     "@types/react": "18.0.9",
     "@types/react-dom": "18.0.4",
     "@types/react-router-dom": "5.3.3",
-    "@types/styled-components": "^5.1.25",
+    "@types/styled-components": "5.1.25",
     "@typescript-eslint/eslint-plugin": "5.23.0",
     "@typescript-eslint/parser": "5.23.0",
     "@vitejs/plugin-react": "1.3.0",
