+++ src/App.tsx
@@ -1,8 +1,8 @@
-import { StrictMode } from 'react'
+import { FC, StrictMode } from 'react'
 import { AppRouter } from './routers/AppRouter'
 import { GlobalStyle } from './styles/GlobalStyles'
 
-const App = () => {
+const App: FC = () => {
   return (
     <StrictMode>
       <GlobalStyle />
