+++ src/const/colors.ts
@@ -9,7 +9,7 @@ export const colors = {
   blue: '#0cdfe8',
   darkGreen: '#4fb48b',
   intenseGreen: '#5df54c',
-  Black: '#060606',
+  black: '#060606',
   white: '#F3EAD3',
   lightBrown: '#DBCCB4',
   darkBrown: '#5C4B30',
