+++ src/pages/LoginScreen/LoginScreen.interface.ts
@@ -1,0 +1,5 @@
+
+export interface FormType {
+  campo: string
+  valido: string
+}
