+++ src/routers/AppRouter.tsx
@@ -1,8 +1,9 @@
+import { FC } from 'react'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import { LoginScreen } from '../pages/LoginScreen'
+import { LoginScreen } from '../pages/LoginScreen/LoginScreen'
 import { DashboardRoutes } from './DashboardRoutes'
 
-export const AppRouter = () => {
+export const AppRouter: FC = () => {
   return (
     <BrowserRouter>
       <Routes>
