+++ src/routers/DashboardRoutes.tsx
@@ -1,4 +1,6 @@
-export const DashboardRoutes = () => {
+import { FC } from 'react'
+
+export const DashboardRoutes: FC = () => {
   return (
     <div>
       <h1>DashboardRoutes pages</h1>
